+++ Assets/Scripts/Core/Mines/MineShape.cs
@@ -1,8 +1,0 @@
-public enum MineShape
-{
-    Single,     // Single cell
-    Cross,      // + shape
-    Square,     // Square shape
-    Diamond,    // Diamond shape
-    Line        // Straight line
-}